--- a/be/src/runtime/data_stream_sender.cpp
+++ b/be/src/runtime/data_stream_sender.cpp
@@ -598,11 +598,13 @@ Status DataStreamSender::send_chunk(RuntimeState* state, vectorized::Chunk* chun
 
             // compute row indexes for each channel
             _channel_row_idx_start_points.assign(num_channels + 1, 0);
+
             for (uint16_t i = 0; i < num_rows; ++i) {
                 uint16_t channel_index = _hash_values[i] % num_channels;
                 _channel_row_idx_start_points[channel_index]++;
                 _hash_values[i] = channel_index;
             }
+
             // NOTE:
             // we make the last item equal with number of rows of this chunk
             for (int i = 1; i <= num_channels; ++i) {