## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
#9118

## Problem Summary(Required) ：
dataset is from tpcds 1T

for such query:
```
mysql>  select  count(*) c, sum(cs_item_sk) from catalog_sales group by cs_order_number order by c limit 10;
```

we need cost about 1sec for caculate shuffle hash, and it's a hotpot in CPU
```
- ShuffleHashTime: 841.121ms
  - __MAX_OF_ShuffleHashTime: 981.712ms
  - __MIN_OF_ShuffleHashTime: 623.501ms
```

perf:
```
  35.12%  starrocks_be        [.] starrocks::Aggregator::build_hash_map<starrocks::vectorized::AggHashMapWithOneNumberKey<(starrocks::PrimitiveType)5, phmap::flat_hash_map<int, un   
  9.16%  starrocks_be        [.] starrocks::vectorized::AggregateFunctionBatchHelper<starrocks::vectorized::AggregateCountFunctionState<false>, starrocks::vectorized::CountAggreg   
  6.94%  starrocks_be        [.] starrocks::pipeline::ExchangeSinkOperator::push_chunk
   3.75%  starrocks_be        [.] starrocks::vectorized::AggregateFunctionBatchHelper<starrocks::vectorized::AggregateCountFunctionState<false>, starrocks::vectorized::CountAggreg   
   3.63%  [vdso]              [.] __vdso_clock_gettime
```

it was because modulo operator break auto vectorized when caculate hash:
```
     │      _ZN9starrocks8HashUtil10xorshift32Ej():                                                                                                                             ▒  
1.80 │        mov    %edi,%edx                                                                                                                                                  ▒  
0.01 │        shl    $0xd,%edx                                                                                                                                                  ▒  
0.01 │        xor    %edi,%edx                                                                                                                                                  ◆  
0.02 │        mov    %edx,%eax                                                                                                                                                  ▒  
1.75 │        shr    $0x11,%eax                                                                                                                                                 ▒  
0.02 │        xor    %eax,%edx                                                                                                                                                  ▒  
0.01 │        mov    %edx,%eax                                                                                                                                                  ▒
     │        shl    $0x5,%eax                                                                                                                                                  ▒  
1.46 │        xor    %edx,%eax                                                                                                                                                  ▒       
     │      _ZN9starrocks8pipeline20ExchangeSinkOperator10push_chunkEPNS_12RuntimeStateERKSt10shared_ptrINS_10vectorized5ChunkEE():                                             ▒  
0.05 │        xor    %edx,%edx                                                                                                                                                  ▒  
0.01 │        div    %r8d                                                                                                                                                       ▒ 
33.46│        mov    %edi,%eax                                                                                                                                                  ▒  
0.17 │        mov    %rdx,%r11                                                                                                                                                  ▒  
0.09 │        xor    %edx,%edx                                                                                                                                                  ▒
     │        div    %r12                                                                                                                                                       ▒
31.64│        imul   %rdx,%r8                                                                                                                                                   ▒
1.25 │        add    %r8,%r11                                                                                                                                                   ▒
0.63 │        mov    %r11d,(%rsi,%r9,4)                                                                                                                                         ▒
3.69 │        add    $0x1,%r9 
```


but we can avoid this operator:
https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/

```
uint32_t reduce(uint32_t x, uint32_t N) {
  return ((uint64_t) x * (uint64_t) N) >> 32 ;
}
```

after this change HashUtil::xorshift32 also could auto-vectorized:

profile:
```
- ShuffleHashTime: 299.518ms
- __MAX_OF_ShuffleHashTime: 347.797ms
- __MIN_OF_ShuffleHashTime: 229.225ms
```
```
  36.65%  starrocks_be        [.] starrocks::Aggregator::build_hash_map<starrocks::vectorized::AggHashMapWithOneNumberKey<(starrocks::PrimitiveType)5, phmap::flat_hash_map<int, u◆
  10.18%  starrocks_be        [.] starrocks::vectorized::AggregateFunctionBatchHelper<starrocks::vectorized::AggregateCountFunctionState<false>, starrocks::vectorized::CountAggre▒
   5.17%  starrocks_be        [.] starrocks::vectorized::AggregateFunctionBatchHelper<starrocks::vectorized::AggregateCountFunctionState<false>, starrocks::vectorized::CountAggre▒   
   4.92%  [vdso]              [.] __vdso_clock_gettime                                                                                                                            ▒   
   2.86%  starrocks_be        [.] starrocks::vectorized::AggregateFunctionBatchHelper<starrocks::vectorized::SumAggregateState<long>, starrocks::vectorized::SumAggregateFunction<▒   
   2.42%  libc-2.17.so        [.] __memset_sse2                                                                                                                                   ▒   
   2.30%  starrocks_be        [.] starrocks::vectorized::AggregateFunctionBatchHelper<starrocks::vectorized::SumAggregateState<long>, starrocks::vectorized::SumAggregateFunction<▒   
   2.26%  libc-2.17.so        [.] __memcpy_ssse3_back                                                                                                                             ▒   
   1.97%  starrocks_be        [.] starrocks::pipeline::ExchangeSinkOperator::push_chunk                                                                                           ▒   
   1.91%  [kernel]            [k] copy_user_enhanced_fast_string                                                                                                                  ▒   
   1.78%  starrocks_be        [.] LZ4_compress_fast_continue                                                                                                                      ▒   
   1.77%  starrocks_be        [.] starrocks::pipeline::PipelineDriverPoller::run_internal                                                                                         ▒   
   1.69%  starrocks_be        [.] starrocks::vectorized::FixedLengthColumnBase<int>::append_selective                                                                             ▒   
   1.33%  libc-2.17.so        [.] __memmove_ssse3_back 
```

baseline: 
5.98 sec
patched:
5.76 sec