@stdpain 2022-07-29
run starrocks_fe_unittest