@stdpain 2022-08-08
run starrocks_be_unittest