@stdpain 2022-08-08
run starrocks_admit_test