@wanpengfei-git 2022-08-09
run starrocks_admit_test